finish tests for getItemIndexByChar
